edit, delete and cancel question functionality completed

+++ JS Applications/Workshop End-to-End Application - Part 1/Quiz System Fever/src/views/edit.js
@@ -1,4 +1,4 @@
-import { deleteQuestion, getQuizQuestions } from "../data/questions.js";
+import { deleteQuestion, getQuestionById, getQuizQuestions } from "../data/questions.js";
 import { getQuizById, getUniqueTopics, updateQuiz } from "../data/quzzes.js";
 import { html, render, page, renderTemplate } from "../lib.js";
 import { createSubmitHandler, getUserData } from "../util.js";
@@ -31,9 +31,9 @@ const editTemplate = (quiz, uniqueTopics, questions, onSaveTitleTopic) => html`
             <h2>Questions</h2>
         </header>
 
-        <div class="pad-large alt-page">
-
-            ${questions.map(question => editDeleteTemplate(question))}
+        <div id="print-questions" class="pad-large alt-page">
+            
+            ${questions.map(question => startingTemplate(question))}
 
             <article class="editor-question">
                 <div class="editor-input">
@@ -49,57 +49,58 @@ const editTemplate = (quiz, uniqueTopics, questions, onSaveTitleTopic) => html`
 </section>
 `;
 
-const editDeleteTemplate = (question) => {
-    countQuestions++;
+const startingTemplate = () => {
     return html`
-    <article class="editor-question">
-                <div class="layout">
-                    <div class="question-control" data-id="${question.objectId}">
-                        <button class="input submit action" @click=${onEditQuestion}><i class="fas fa-edit"></i> Edit</button>
-                        <button class="input submit action" @click=${onDeleteQuestion}><i class="fas fa-trash-alt"></i> Delete</button>
-                    </div>
-                    <h3>Question ${countQuestions}</h3>
-                </div>
-                <form>
-                    <p class="editor-input">${question.text}</p>
-                    <div class="editor-input">
-                        <label class="radio">
-                            <input class="input" type="radio" name="question" value="0" disabled />
-                            <i class="fas fa-check-circle"></i>
-                        </label>
-                        <span>${question.answers[0]}</span>
-                    </div>
-                    <div class="editor-input">
-                        <label class="radio">
-                            <input class="input" type="radio" name="question" value="1" disabled />
-                            <i class="fas fa-check-circle"></i>
-                        </label>
-                        <span>${question.answers[1]}</span>
-                    </div>
-                    <div class="editor-input">
-                        <label class="radio">
-                            <input class="input" type="radio" name="question" value="2" disabled />
-                            <i class="fas fa-check-circle"></i>
-                        </label>
-                        <span>${question.answers[2]}</span>
-                    </div>
-                </form>
+            <article class="editor-question">
             </article> `;
 }
 
-const saveCancelTemplate = () => html`
-    <article class="editor-question">
+const editDeleteTemplate = (question, questionNum) => html`
+    <div class="layout">
+            <div class="question-control" data-id="${question.objectId}">
+                <button class="input submit action" @click=${onEditQuestion} data-id="${questionNum}"><i class="fas fa-edit"></i> Edit</button>
+                <button class="input submit action" @click=${onDeleteQuestion}><i class="fas fa-trash-alt"></i> Delete</button>
+            </div>
+            <h3>Question ${questionNum}</h3>
+        </div>
+        <form>
+            <p class="editor-input">${question.text}</p>
+            <div class="editor-input">
+                <label class="radio">
+                    <input class="input" type="radio" name="question" value="0" disabled />
+                    <i class="fas fa-check-circle"></i>
+                </label>
+                <span>${question.answers[0]}</span>
+            </div>
+            <div class="editor-input">
+                <label class="radio">
+                    <input class="input" type="radio" name="question" value="1" disabled />
+                    <i class="fas fa-check-circle"></i>
+                </label>
+                <span>${question.answers[1]}</span>
+            </div>
+            <div class="editor-input">
+                <label class="radio">
+                    <input class="input" type="radio" name="question" value="2" disabled />
+                    <i class="fas fa-check-circle"></i>
+                </label>
+                <span>${question.answers[2]}</span>
+            </div>
+        </form>
+`;
+
+const saveCancelTemplate = (question, questionNum) => html`
                 <div class="layout">
-                    <div class="question-control">
-                        <button class="input submit action"><i class="fas fa-check-double"></i>
+                    <div class="question-control" data-id="${question.objectId}">
+                        <button class="input submit action" @click=${onSaveQuestion}><i class="fas fa-check-double"></i>
                             Save</button>
-                        <button class="input submit action"><i class="fas fa-times"></i> Cancel</button>
+                        <button class="input submit action" @click=${onCancelChanges} data-id="${questionNum}"><i class="fas fa-times"></i> Cancel</button>
                     </div>
-                    <h3>Question 1</h3>
+                    <h3>Question ${questionNum}</h3>
                 </div>
                 <form>
                     <textarea class="input editor-input editor-text" name="text"
-                        placeholder="Enter question"></textarea>
+                        placeholder="Enter question" .value=${question.text}></textarea>
                     <div class="editor-input">
 
                         <label class="radio">
@@ -107,7 +108,7 @@ const saveCancelTemplate = () => html`
                             <i class="fas fa-check-circle"></i>
                         </label>
 
-                        <input class="input" type="text" name="answer-0" />
+                        <input class="input" type="text" name="answer-0" .value="${question.answers[0]}" />
                         <button class="input submit action"><i class="fas fa-trash-alt"></i></button>
                     </div>
                     <div class="editor-input">
@@ -117,7 +118,7 @@ const saveCancelTemplate = () => html`
                             <i class="fas fa-check-circle"></i>
                         </label>
 
-                        <input class="input" type="text" name="answer-1" />
+                        <input class="input" type="text" name="answer-1" .value="${question.answers[1]}" />
                         <button class="input submit action"><i class="fas fa-trash-alt"></i></button>
                     </div>
                     <div class="editor-input">
@@ -127,7 +128,7 @@ const saveCancelTemplate = () => html`
                             <i class="fas fa-check-circle"></i>
                         </label>
 
-                        <input class="input" type="text" name="answer-2" />
+                        <input class="input" type="text" name="answer-2" .value="${question.answers[2]}" />
                         <button class="input submit action"><i class="fas fa-trash-alt"></i></button>
                     </div>
                     <div class="editor-input">
@@ -137,7 +138,6 @@ const saveCancelTemplate = () => html`
                         </button>
                     </div>
                 </form>
-            </article>
 `;
 
 const cannotEditTemplate = () => html`
@@ -208,6 +208,16 @@ export async function showEdit(ctx) {
     const questions = await getQuizQuestions(quizId);
 
     render(editTemplate(quiz, uniqueTopics.results, questions.results, createSubmitHandler(onSaveTitleTopic)));
+    renderQuestions(questions.results);
+}
+
+function renderQuestions(questions) {
+    const articles = document.getElementById('print-questions').querySelectorAll('article');
+
+    for (let i = 0; i < questions.length; i++) {
+        countQuestions++;
+        renderTemplate(editDeleteTemplate(questions[i], countQuestions), articles[i]);
+    }
 }
 
 async function onSaveTitleTopic({ title, topic }, form) {
@@ -233,23 +243,49 @@ async function onSaveTitleTopic({ title, topic }, form) {
     form.reset();
 }
 
-async function onEditQuestion() {
+async function onEditQuestion(e) {
+    if (e.target.tagName == 'BUTTON') {
+        const questionId = e.target.parentElement.dataset.id;
+        const question = await getQuestionById(questionId);
+        const questionNum = e.target.dataset.id;
 
+        const root = e.target.parentElement.parentElement.parentElement;
+        renderTemplate(saveCancelTemplate(question, questionNum), root);
+    }
 }
 
 async function onDeleteQuestion(e) {
-    const choice = confirm('Are you sure?');
+    if (e.target.tagName == 'BUTTON') {
+        const choice = confirm('Are you sure?');
 
-    if (choice) {
-        const questionId = e.target.parentElement.dataset.id;
-        await deleteQuestion(questionId);
+        if (choice) {
+            const questionId = e.target.parentElement.dataset.id;
+            await deleteQuestion(questionId);
 
-        const article = e.target.parentElement.parentElement.parentElement;
-        const divOverlay = document.createElement('div');
-        divOverlay.classList.add('loading-overlay', 'working');
-        article.appendChild(divOverlay);
+            const article = e.target.parentElement.parentElement.parentElement;
+            const divOverlay = document.createElement('div');
+            divOverlay.classList.add('loading-overlay', 'working');
+            article.appendChild(divOverlay);
+        }
     }
+}
+
+async function onSaveQuestion(e) {
+    if (e.target.tagName == 'BUTTON') {
 
 
+
+    }
+}
+
+async function onCancelChanges(e) {
+    if (e.target.tagName == 'BUTTON') {
+        const questionId = e.target.parentElement.dataset.id;
+        const question = await getQuestionById(questionId);
+        const questionNum = e.target.dataset.id;
+
+        const root = e.target.parentElement.parentElement.parentElement;
+        renderTemplate(editDeleteTemplate(question, questionNum), root);
+    }
 }
